For the login of Kaggle, I changed the logging API key with the user doing it manually rather than it being in the code.
After looking at the results from the other models I decided to do an embed model for the 
neural net and the boost. iv retained the xgboost model with hyper tuning and trained the neural net again
iv then combined them into an embeded structure then evaluated it on the test dataset.